polished Show page. developed the rest of the user managment functionality. added some authorization checks.

+++ controllers/lognotes.js
@@ -11,27 +11,36 @@ const User = require('../models/user.js');
 // import Lognote model
 const Lognote = require('../models/lognote.js');
 
-// -------------------------------------------------------------- Routes
+// -------------------------------------------------------------- r o u t e s
 
 // -------------------------------------------------------------- Index
 
 router.get('/', async (req, res) => {
-    
-    // add logic to check if logged in
-
-    const lognotes = await Lognote.find();
-
-    res.render('lognotes/index.ejs', {
-        lognotes: lognotes,
-    });
+    if (req.session.user) {
+        const lognotes = await Lognote.find();
+        res.render('lognotes/index.ejs', {
+            lognotes: lognotes,
+        });
+    } else {
+        res.redirect('/auth/sign-in.ejs');
+    };
 });
 
 // -------------------------------------------------------------- New
 
-router.get('/new', (req, res) => {
-    
-    // add logic to check if logged in
-    res.render('lognotes/new.ejs');
+router.get('/new', async (req, res) => {
+    if (req.session.user) {
+        const currentUser = await User.findById(req.session.user._id);
+        if (currentUser.role === 'admin' || currentUser.role === 'editor') {
+            res.render('lognotes/new.ejs', {
+                user: currentUser,
+            });
+        } else {
+            res.redirect('/');
+        };
+    } else {
+        res.redirect('/auth/sign-in.ejs');
+    };
 });
 
 // -------------------------------------------------------------- Delete
@@ -39,8 +48,13 @@ router.get('/new', (req, res) => {
 router.delete('/:lognoteId', async (req, res) => {
     try {
         // this should have some sort of confirmation double-check
-        await Lognote.findByIdAndDelete(req.params.lognoteId);
+
+        const currentUser = await User.findById(req.session.user._id);
+        if (currentUser.role === 'admin' || currentUser.role === 'editor') {
+            await Lognote.findByIdAndDelete(req.params.lognoteId);
+        };
         res.redirect('/lognotes');
+
     } catch (error) {
         console.log(error);
         res.redirect('/');
@@ -50,17 +64,16 @@ router.delete('/:lognoteId', async (req, res) => {
 // -------------------------------------------------------------- Update
 
 router.put('/:lognoteId', async (req, res) => {
-    console.log(req.body);
     try {
-        
-        const lognote = await Lognote.findById(req.params.lognoteId);
-
-        lognote.set(req.body);
-
-        await lognote.save();
-        
-        res.redirect('/lognotes');
-
+        const currentUser = await User.findById(req.session.user._id);
+        if (currentUser.role === 'admin' || currentUser.role === 'editor') {
+            const lognote = await Lognote.findById(req.params.lognoteId);
+            lognote.set(req.body);
+            await lognote.save();
+            res.redirect('/lognotes');
+        } else {
+            res.redirect('/');
+        };
     } catch (error) {
         console.log(error);
         res.redirect('/');
@@ -70,21 +83,22 @@ router.put('/:lognoteId', async (req, res) => {
 // -------------------------------------------------------------- Create
 
 router.post('/', async (req, res) => {
-
     try {
-
         // auto-fill the creator
-        // const currentUser = await User.findById(req.session.user._id);
 
         // auto-generate the created timestamp
-
 
         // turn check boxes into true/false
         // parse dates & times
 
-        await Lognote.create(req.body);
+        const currentUser = await User.findById(req.session.user._id);
 
-        res.redirect('/lognotes');
+        if (currentUser.role === 'admin' || currentUser.role === 'editor') {
+            await Lognote.create(req.body);
+            res.redirect('/lognotes');
+        } else {
+            res.redirect('/');
+        };
     } catch (error) {
         console.log(error);
         res.redirect('/');
@@ -94,38 +108,51 @@ router.post('/', async (req, res) => {
 // -------------------------------------------------------------- Edit
 
 router.get('/:lognoteId/edit', async (req, res) => {
+    if (req.session.user) {
+        try {
 
-    try {
+            const currentUser = await User.findById(req.session.user._id);
+            if (currentUser.role === 'admin' || currentUser.role === 'editor') {
 
-        const lognote = await Lognote.findById(req.params.lognoteId);
-        
-        res.render('lognotes/edit.ejs', {
-            lognote: lognote,
-        });
+                const lognote = await Lognote.findById(req.params.lognoteId);
+                res.render('lognotes/edit.ejs', {
+                    lognote: lognote,
+                });
 
-    } catch (error) {
-        console.log(error);
-        res.redirect('/');
+            } else {
+                res.redirect('/');
+            };
+
+        } catch (error) {
+            console.log(error);
+            res.redirect('/');
+        };
+    } else {
+        // start at the sign in page
+        res.redirect('/auth/sign-in.ejs');
     };
 });
 
 // -------------------------------------------------------------- Show
 
 router.get('/:lognoteId', async (req, res) => {
+    if (req.session.user) {
+        try {
+            const lognote = await Lognote.findById(req.params.lognoteId);
 
-    try {
-        const lognote = await Lognote.findById(req.params.lognoteId);
-        
-        res.render('lognotes/show.ejs', {
-            lognote: lognote,
-        });
+            res.render('lognotes/show.ejs', {
+                lognote: lognote,
+            });
 
-    } catch (error) {
-        console.log(error);
-        res.redirect('/');
+        } catch (error) {
+            console.log(error);
+            res.redirect('/');
+        };
+    } else {
+        // start at the sign in page
+        res.redirect('/auth/sign-in.ejs');
     };
 });
-
 
 // -------------------------------------------------------------- export router
